folder restructure for user widget
