--- docs/_includes/arguments/icon-button-list.md
@@ -1,7 +1,0 @@
-| Name          | Type    | Required | Description                               |
-| ------------- | ------- | -------- | ----------------------------------------- |
-| disabled      | boolean | Yes      | Defines whether the button is disabled    |
-| id            | string  | Yes      | The button ID                             |
-| icon          | string  | Yes      | The icon name. See Static Icons for names |
-| tooltipText   | string  | Yes      | The text to show in the tooltip           |
-| ariaLabelText | string  | Yes      | The text to show int the screenreaders    |
