--- docs/_includes/arguments/static-icon.md
@@ -1,4 +1,0 @@
-| Name | Type   | Required | Description                                    |
-| ---- | ------ | -------- | ---------------------------------------------- |
-| name | object | Yes      | The icon name                                  |
-| size | string | no       | The size of the icon - valid values: 'l', 'xl' |
