--- docs/components/icon-button-list.md
@@ -1,10 +1,0 @@
----
-layout: layouts/component.njk
-title: Icon Button List
----
-
-{% example "icon-button-list", 300 %}
-
-## When to use
-
-Use this component when want a set as buttons represented as icons
