--- docs/components/static-icon.md
@@ -1,10 +1,0 @@
----
-layout: layouts/component.njk
-title: Static Icon
----
-
-{% example "static-icon", 300 %}
-
-## When to use
-
-Use this component when you need an icon
