--- docs/components/table-card-group.md
@@ -1,10 +1,0 @@
----
-layout: layouts/component.njk
-title: Table Card Group
----
-
-{% example "table-card-group", 300 %}
-
-{% header 2, "When to use" %}
-
-Use this component when users need to represent a list of items as a list of cards, and within a table, and the ability to toggle between the two.
