--- docs/examples/icon-button-list/index.njk
@@ -1,31 +1,0 @@
----
-layout: layouts/example.njk
-title: Icon Button List (example)
-arguments: icon-button-list
----
-
-{% from "dpr/components/icon-button-list/view.njk" import dprIconButtonList %}
-
-{{ dprIconButtonList([
-  {
-    disabled: false,
-    id: 'button-id-1',
-    icon: 'print',
-    tooltipText: 'tooltip text',
-    ariaLabelText: 'tooltip text'
-  },
-  {
-    disabled: true,
-    id: 'button-id-1',
-    icon: 'copy',
-    tooltipText: 'tooltip text',
-    ariaLabelText: 'tooltip text'
-  },
-  {
-    disabled: false,
-    id: 'button-id-3',
-    icon: 'share',
-    tooltipText: 'tooltip text',
-    ariaLabelText: 'tooltip text'
-  }
-]) }}
