--- docs/examples/static-icon/index.njk
@@ -1,62 +1,0 @@
----
-layout: layouts/example.njk
-title: Static Icon (example)
-arguments: static-icon
----
-
-{% from "dpr/components/static-icon/view.njk" import dprStaticIcon %}
-
-<p>bookmark</p>
-{{ dprStaticIcon('bookmark', '') }}
-{{ dprStaticIcon('bookmark', 'l') }}
-{{ dprStaticIcon('bookmark', 'xl') }}
-
-<p>viewed</p>
-{{ dprStaticIcon('viewed', '') }}
-{{ dprStaticIcon('viewed', 'l') }}
-{{ dprStaticIcon('viewed', 'xl') }}
-
-<p>list</p>
-{{ dprStaticIcon('list', '') }}
-{{ dprStaticIcon('list', 'l') }}
-{{ dprStaticIcon('list', 'xl') }}
-
-<p>card</p>
-{{ dprStaticIcon('card', '') }}
-{{ dprStaticIcon('card', 'l') }}
-{{ dprStaticIcon('card', 'xl') }}
-
-<p>print</p>
-{{ dprStaticIcon('print', '') }}
-{{ dprStaticIcon('print', 'l') }}
-{{ dprStaticIcon('print', 'xl') }}
-
-<p>download</p>
-{{ dprStaticIcon('download', '') }}
-{{ dprStaticIcon('download', 'l') }}
-{{ dprStaticIcon('download', 'xl') }}
-
-<p>share</p>
-{{ dprStaticIcon('share', '') }}
-{{ dprStaticIcon('share', 'l') }}
-{{ dprStaticIcon('share', 'xl') }}
-
-<p>close</p>
-{{ dprStaticIcon('close', '') }}
-{{ dprStaticIcon('close', 'l') }}
-{{ dprStaticIcon('close', 'xl') }}
-
-<p>info</p>
-{{ dprStaticIcon('info', '') }}
-{{ dprStaticIcon('info', 'l') }}
-{{ dprStaticIcon('info', 'xl') }}
-
-<p>hourglass</p>
-{{ dprStaticIcon('hourglass', '') }}
-{{ dprStaticIcon('hourglass', 'l') }}
-{{ dprStaticIcon('hourglass', 'xl') }}
-
-<p>copy</p>
-{{ dprStaticIcon('copy', '') }}
-{{ dprStaticIcon('copy', 'l') }}
-{{ dprStaticIcon('copy', 'xl') }}
