--- src/dpr/components/async/applied-filters/style.scss
@@ -1,27 +1,0 @@
-.async-applied-filters-list {
-  padding-top: govuk-spacing(3);
-  padding-bottom: govuk-spacing(3);
-
-  .async-applied-filters-list__item {
-    display: inline-block;
-    margin-right: 10px;
-    p {
-      margin-bottom: 0px;
-    }
-  }
-
-  .async-applied-filters-list__item--title {
-  } 
-}
-
-@media print {
-  .async-applied-filters-list {
-    .async-applied-filters-list__item {
-      display: block;
-      p {
-        font-size: 14px;
-      }
-    }
-  }
-}
-
