--- src/dpr/components/async/applied-filters/view.njk
@@ -1,12 +1,0 @@
-{% macro dprAsyncAppliedFiltersList(filtersList) %}
-  <div class="async-applied-filters-list">
-    <div class="async-applied-filters-list__item async-applied-filters-list__item--title">
-      <p class="govuk-heading-s">Filters:</p>
-    </div>
-    {% for filter in filtersList %}
-      <div class="async-applied-filters-list__item">
-        <p class="govuk-body-s">{{filter.name}}: {{ filter.value }}</p>
-      </div>
-    {% endfor %}
-  </div>
-{% endmacro %}
