--- src/dpr/components/content-slide-cards/style.scss
@@ -1,41 +1,0 @@
-.dpr-card-group {
-  padding-bottom: govuk-spacing(6);
-  .dpr-card-group__item {
-    padding-bottom: govuk-spacing(5);
-    position: relative;
-
-    .dpr-card-group__item-header {
-      margin-bottom: govuk-spacing(2);
-    }
-
-    .dpr-card-group__item__description {
-      margin-top: 10px;
-    }
-
-    .dpr-card-group__item__timestamp {
-      position: absolute;
-      bottom: 1px;
-      left: 0;
-      padding-bottom: 0;
-      margin-bottom: 0;
-      font-size: 14px;
-      color: #505a5f;
-    }
-
-    .dpr-card-group__item__status {
-      position: absolute;
-      bottom: 0;
-      right: 0;
-    }
-  }
-}
-
-ul.dpr-card-group__item__filters-list {
-  list-style-type: none;
-  padding: 0;
-  margin-top: 0px;
-
-  li {
-    margin-bottom: 0;
-  }
-}
