--- src/dpr/components/content-slide-cards/view.njk
@@ -1,54 +1,0 @@
-{% from "../request-status-tag/view.njk" import dprRequestStatusTag %}
-{% from "govuk/components/details/macro.njk" import govukDetails %}
-
-{% macro dprContentSlideCards(items) %}
-  <div class="dpr-card-group">
-    {% for item in items %}
-      {% set status = item.status %}
-      {% set timestamp = item.timestamp %}
-      {% set type = item.type %}
-      {% set executionId = item.executionId %}
-
-      <div class="dpr-card-group__item" data-href="{{ item.href }}">
-        <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible govuk-!-static-margin-top-0">
-
-        {# HEADING #}
-        <div class="dpr-card-group__item-header">
-          <a class="dpr-content-card__item__title govuk-link govuk-!-font-weight-bold govuk-link--no-visited-state govuk-body card__link" href="{{ item.href }}">
-            {{ item.text }}
-          </a>
-        </div>
-
-        {# DESCRIPTION / FILTERS #}
-        {{ govukDetails({
-          summaryText: 'Description',
-          text: item.description,
-          classes: 'dpr-meta-data-details'
-        }) }}
-
-        {% if item.summary %}
-          <ul class="dpr-card-group__item__filters-list">
-            {% for filter in item.summary %}
-              <li class="govuk-body-s">{{ filter.name }}: {{ filter.value }}</li>
-            {% endfor %}
-            {% for sorted in item.sort %}
-              <li class="govuk-body-s">{{ sorted.name }} : {{ sorted.value }}</li>
-            {% endfor %}
-          </ul>
-        {% endif %}
-
-        {# TIMESTAMP #}
-        {% if timestamp %}
-          <p class='dpr-card-group__item__timestamp'>{{ timestamp }}</p>
-        {% endif %}
-
-        {# STATUS #}
-        {% if status %}
-          <div class="dpr-card-group__item__status">
-            {{ dprRequestStatusTag(status) }}
-          </div>
-        {% endif %}
-      </div>
-    {% endfor %}
-  </div>
-{% endmacro %}
