--- src/dpr/components/content-slide/view.njk
@@ -1,44 +1,0 @@
-{% from "../static-icon/view.njk" import dprStaticIcon %}
-{% from "../toggle-content/view.njk" import dprToggleContent %}
-
-{% macro dprContentSlide(head, content, total) %}
-
-  {# HEADING CONTENT #}
-  {% set title = head.title %}
-  {% set subText = head.subText %}
-  {% set icon = head.icon %}
-  {% set id = head.id %}
-  {% set emptyMessage = head.emptyMessage %}
-  {% set href = head.href %}
-
-  {# TOGGLE CONTENT #}
-  {% set toggle = content.toggle %}
-  {% set toggleIcons = content.icons %}
-  {% set items = content.items %}
-
-  <div class="dpr-slide-container">
-    <div class="dpr-slide__header">
-      <h2 class="govuk-heading-m dpr-slide__title">
-        {% if href %}
-          <a href='{{ href }}'>{{ title }}</a>
-        {% else %}
-          {{ title }}
-        {% endif %}
-      </h2>
-
-      {% if not emptyMessage.length and total %}
-        <p class="govuk-body dpr-slide__sub-text">Showing {{ total.shown }} of  {{ total.amount }}. </p>
-      {% endif %}
-    </div>
-
-    {% if emptyMessage.length %}
-      <p>{{emptyMessage}}</p>
-    {% else %}
-      {% if items and items.length > 1 %}
-        {{ dprToggleContent( id, toggleIcons, items) }}
-      {% else %}
-        {{ items[0].html | safe }}
-      {% endif %}
-    {% endif %}
-  </div>
-{% endmacro %}
