--- src/dpr/components/icon-button-list/style.scss
@@ -1,19 +1,0 @@
-.icon-button-list {
-  position: absolute;
-  right: 0px;
-  top: 0px;
-
-  .dpr-icon-wrapper {
-    margin-right: 0px;
-
-    &:focus {
-      background-color: govuk-colour("yellow");
-    }
-  }
-}
-
-@media print {
-  .icon-button-list {
-    display: none;
-  }
-}
