--- src/dpr/components/icon-button-list/view.njk
@@ -1,17 +1,0 @@
-{% macro dprIconButtonList(buttonList) %}
-  <div class="icon-button-list" data-dpr-module='icon-button-list'>
-    {% for button in buttonList %}
-      <div
-        tabindex="0"
-        id="{{ button.id }}"
-        data-href="{{ button.href }}"
-        class="dpr-icon-wrapper dpr-icon-wrapper--l dpr-icon-live {% if button.disabled %}dpr-icon-disabled{% endif %}" 
-        tooltip="{{ button.tooltipText }}{% if button.disabled %} is disabled{% endif %}"
-        aria-label="{{ button.ariaLabelText }}{% if button.disabled %} is disabled{% endif %}"
-        data-disabled="{{ button.disabled }}"
-        role="button">
-        <i class="dpr-icon {{ button.icon }}-icon"></i>
-      </div>
-    {% endfor %}
-  </div>
-{% endmacro %}
