+++ src/dpr/components/report-list/view.njk
@@ -5,11 +5,10 @@
   {% from "dpr/components/filters/view.njk" import dprFilters %}
   {% from "dpr/components/loading-anim/view.njk" import dprLoadingAnim -%}
   {% from "dpr/components/column-section/view.njk" import dprColumnSection %}
-  {% from "dpr/components/icon-button-list/view.njk" import dprIconButtonList %}
   {% from "dpr/components/report-actions/view.njk" import dprReportActions %}
   {% from "dpr/components/pagination/view.njk" import dprPagination %}
   {% from "dpr/components/data-table/view.njk" import dprDataTable %}
-  {% from "../report-totals/view.njk" import dprReportTotals %}
+  {% from "dpr/components/report-totals/view.njk" import dprReportTotals %}
   
   {% set title = options.title %}
   {% set dataTableOptions = options.dataTableOptions %}
