--- src/dpr/components/table-card-group/styles.scss
@@ -1,15 +1,0 @@
-.dpr-req-filters-summary {
-  min-width: 200px;
-}
-
-.dpr-toggle-content {
-  .govuk-table__cell {
-    .govuk-details {
-      margin-bottom: 0px;
-    }
-  }
-}
-
-.dpr-description-head {
-  width: 65%;
-}
