--- src/dpr/components/table-card-group/view.njk
@@ -1,39 +1,0 @@
-{% from "../content-slide/view.njk" import dprContentSlide %}
-{% from "../content-slide-cards/view.njk" import dprContentSlideCards %}
-{% from "../reports-list/view.njk" import dprReportsList %}
-{% from "govuk/components/table/macro.njk" import govukTable %}
-
-{% macro dprTableCardGroup(head, cardData, tableData, total, search = false) %}
-
-  {% set icons = [] %}
-  {% set items = [] %}
-
-  {% if tableData %}
-    {% if search %}
-      {% set tableHtml %}
-      {{ dprReportsList(tableData.head, tableData.rows) }}
-      {% endset %}
-    {% else %}
-      {% set tableHtml %}
-      {{ govukTable(tableData) }}
-      {% endset %}
-    {% endif %}
-
-    {% set icons = (icons.push({name: 'list', tooltip: 'List view', active: true}), icons) %}
-    {% set items = (items.push({active: true, html: tableHtml}), items) %}
-  {% endif %}
-
-  {% if cardData %}
-    {% set cardHtml %}
-    {{ dprContentSlideCards(cardData) }}
-    {% endset %}
-
-    {% set icons = (icons.push({name: 'card', tooltip: 'Card view', active: false}), icons) %}
-    {% set items = (items.push({active: false, html: cardHtml}), items) %}
-  {% endif %}
-
-  {% set content = { icons: icons, items: items }%}
-
-  {{ dprContentSlide(head, content, total) }}
-
-{% endmacro %}
