+++ src/dpr/components/user-reports-bookmarks-list/view.njk
@@ -1,0 +1,7 @@
+{% from "../user-reports-list/view.njk" import dprUserReportsList %}
+
+{% macro dprBookmarksList(args) %}
+  <div id="dpr-bookmarks-list">
+    {{ dprUserReportsList(args) }}
+  </div>
+{% endmacro %}
