+++ src/dpr/components/user-reports-lists/view.njk
@@ -1,6 +1,7 @@
-{% from "../user-list-content/view.njk" import dprUserList %}
-{% from "../async-request-list/view.njk" import dprAsyncRequestList %}
-{% from "../recently-viewed-list/view.njk" import dprRecentlyViewedList %}
+{% from "../user-reports-list/view.njk" import dprUserReportsList %}
+{% from "../user-reports-request-list/view.njk" import dprAsyncRequestList %}
+{% from "../user-reports-viewed-list/view.njk" import dprRecentlyViewedList %}
+{% from "../user-reports-bookmarks-list/view.njk" import dprBookmarksList %}
 {% from "govuk/components/tabs/macro.njk" import govukTabs %}
 
 {% macro dprUserReportsLists(args) %}
@@ -52,13 +53,11 @@
 
     {% if bookmarks %}
       {% set bookmarksHtml %}
-        <div id="dpr-bookmarks-list">
-          {{ dprUserList({
-            head: bookmarks.head,
-            tableData: bookmarks.tableData,
-            total: bookmarks.total
-          }) }}
-        </div>
+        {{ dprBookmarksList({
+          head: bookmarks.head,
+          tableData: bookmarks.tableData,
+          total: bookmarks.total
+        }) }}
       {% endset -%}
       
       {% set items = (items.push({
