+++ src/dpr/components/user-reports-viewed-list/clientClass.mjs
