+++ src/dpr/components/user-reports-viewed-list/view.njk
@@ -1,7 +1,7 @@
-{% from "../user-list-content/view.njk" import dprUserList %}
+{% from "../user-reports-list/view.njk" import dprUserReportsList %}
 
 {% macro dprRecentlyViewedList(args) %}
   <div id="dpr-recently-viewed-component" data-dpr-module="recently-viewed-list" data-request-data='{{ args.meta | dump | safe }}' data-csrf-token="{{ args.csrfToken }}">
-    {{ dprUserList(args) }}
+    {{ dprUserReportsList(args) }}
   </div>
 {% endmacro %}
