+++ test-app/async.njk
@@ -1,14 +1,6 @@
 {% extends "page.njk" %}
 {% from "govuk/components/breadcrumbs/macro.njk" import govukBreadcrumbs %}
-{% from "govuk/components/tabs/macro.njk" import govukTabs %}
-{% from "components/content-slide/view.njk" import dprContentSlide %}
-{% from "components/content-slide-cards/view.njk" import dprContentSlideCards %}
-{% from "components/table-card-group/view.njk" import dprTableCardGroup %}
-{% from "components/async-request-list/view.njk" import dprAsyncRequestList %}
-{% from "components/recently-viewed-list/view.njk" import dprRecentlyViewedList %}
 {% from "components/reports-list/view.njk" import dprReportsList %}
-{% from "components/tag/view.njk" import dprTag %}
-{% from "components/user-list-content/view.njk" import dprUserList %}
 {% from "components/user-reports-lists/view.njk" import dprUserReportsLists %}
 
 {% block content %}
