+++ test-app/server.js
@@ -15,8 +15,8 @@ const bodyParser = require('body-parser')
 const { default: reportListUtils } = require('../package/dpr/components/report-list/utils')
 const BookmarklistUtils = require('../package/dpr/utils/bookmarkListUtils').default
 const ReportslistUtils = require('../package/dpr/components/reports-list/utils').default
-const AsyncReportListUtils = require('../package/dpr/components/async-request-list/utils').default
-const RecentlyViewedCardGroupUtils = require('../package/dpr/components/recently-viewed-list/utils').default
+const AsyncReportListUtils = require('../package/dpr/components/user-reports-request-list/utils').default
+const RecentlyViewedCardGroupUtils = require('../package/dpr/components/user-reports-viewed-list/utils').default
 const UserReportsListUtils = require('../package/dpr/components/user-reports-lists/utils').default
 
 // Set up application
